Revert "Enable ccache build option in all configs (#68)"
This reverts commit 3ed61b8dfc3d261a801f5098c5e1bc4ad353b9db.

python distutils does not play well with ccache. Specifically, enabling
ccache caused bad shared libraries for numpy to be built. I cannot pin
down the exact problem. It seems to work correctly on the latest version
of buildroot ("next" branch) as of 2019-06-06. So, we can revisit this
after upgrading to a newer release of buildroot in the future.